Efficiency of optimal design

This function computes the efficiency of the D, I, and A optimal designs compared to the random design.

efficiency(
  yyy,
  ip,
  uncert = FALSE,
  ipop,
  oc = "D",
  L = NULL,
  items = FALSE,
  integ = TRUE
)

Arguments

yyy

a optical object; the output of a call optical()

ip

matrix with item parameters for all items (number of rows determines number of items; number of columns is 2 (for 2PL; or 1PL with common a-parameter when NA in first column from second item) or 3 (for 3PL; or mixed 2/3-PL with NA for 2PL-items in third column)

uncert

if false (default), abilities are assumed to be known; if true, handling of uncertainties of Bjermo et al. (2021) is used.

ipop

matrix with item parameters for operational items (used if uncert=TRUE, only).

oc

optimality criterion: "D" (D-optimality, default), "I" (I-optimality with standard normal weight function), "A" (A-optimality).

L

L-matrix (not used for D-optimality)

items

if false (default), only total block efficiency is returned; if true, criteria for optimal and random and the efficiency for each item are reported in each column of output. Last column are then total criteria and efficiency. D-, L-, I-, A-optimality

integ

if true (default), integrate() is used for computation of partial information matrices; if false, Riemann rule is used.

See also

Examples

# 1PL-models with common discrimination parameter
ip <- cbind(c(1.6, NA), c(-1, 1))

yyy <- optical(ip, oc="D", uncert=FALSE, ipop,
               imf=c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
               nnn=c(0, 50, 50, 200, 200, 200),
               nsp=c(0.001, 0.0001, 0.0001, 0.00001, 0.00001, 0.00001),
               sss=0.001, falpha=1.08, ig=3, ex=0)
#> ---> Outer iteration = 1 
#> ++++++++++++++++++++
#> ---> Adapt grid; outer iteration = 2 
#> ++

# Efficiency of D-optimal design compared to random design
efficiency(yyy, ip, oc="D")
#> [1] 1.237897

# Efficiency of I-optimal design compared to random design
efficiency(yyy, ip, oc="I")
#> [1] 1.086161

# Efficiency of A-optimal design compared to random design
efficiency(yyy, ip, oc="A")
#> [1] 1.086161