Optimal item calibration

Calibrate items following a 2PL, 3PL, mixture of 2PL and 3PL model, or 2PL with common discrimination for all items (Rasch-type).

optical(
  ip,
  oc = "D",
  uncert = FALSE,
  ipop,
  imf = c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
  maxiter = rep(300, 6),
  eps = rep(0.002, 6),
  nnn = c(0, 50, 50, 200, 200, 200),
  nsp = c(0.001, 1e-04, 1e-04, 1e-05, 1e-05, 1e-05),
  sss = 0.001,
  falpha = 1.08,
  sdr = TRUE,
  ig = 3,
  ex = 0,
  integ = TRUE,
  show_progress = FALSE
)

Arguments

ip

matrix with item parameters for all items (number of rows determines number of items number of column 2 (2PL or Rasch-type with NA from second item in first column) or 3 (3PL or mixed 2/3-PL with NA for 2PL-items in third column).

oc

optimality criterion: "D" (D-optimality, default), "I" (I-optimality with standard normal weight function), "A" (A-optimality).

uncert

if false (default), abilities are assumed to be known; if true, handling of uncertainties of Bjermo et al. (2021) is used.

ipop

matrix with item parameters for operational items (used if uncert=TRUE, only).

imf

the vector of step-lengths; default c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45).

maxiter

maximal number of iterations in each inner loop, the length of this vector defines the number of outer loops.

eps

convergence criterion (maximum violation of eq.th.), vector with value for each iteration in the outer loop, but the same number for all iterations is recommended.

nnn

number of new nodes added at each position in the adaptive grid, vector with value for each iteration in the outer loop (nnn [1] not used).

nsp

node spacing between new nodes, vector with value for each iteration in the outer loop (nsp [1] is the spacing between nodes of the starting grid).

sss

step size stopping criterion.

falpha

factor alpha for adjusting the step size vector (should be > 1).

sdr

stop if design repeated (flag TRUE/FALSE).

ig

inner grid between -ig and ig.

ex

intervals of size < ex will be removed (consolidate); if ex=0, no consolidation will be done.

integ

if true (default), integrate() is used for computation of partial information matrices; if false, Riemann rule is used.

show_progress

FALSE (default), This ensures that the + symbols are printed on a line for each Iteration otherwise some output of the function is printed.

Value

Result of this function is a list with following instances:

dd

directional derivatives of optimal solution.

xi

optimal solution.

t

final grid of ability values which was used.

viomax

largest violation of eq.th. from final solution (if < eps, alg. has converged, otherwise not).

h1

interval boundaries for optimal solution.

mooiter

monitoring iterations; information about each iteration to produce convergence plots.

time

running time of algorithm in minutes.

oc

optimality criterion ("D", "I", "A", "L").

L

L-matrix (not for D-optimality).

References

Ul Hassan and Miller (2021). An exchange algorithm for optimal calibration of items in computerized achievement tests.Computational Statistics and Data Analysis, 157: 107177. doi:10.1016/j.csda.2021.107177

Bjermo, Fackle-Fornius, and Miller (2021). Optimizing Calibration Designs with Uncertaintyin Abilities. Manuscript. urn:nbn:se:su:diva-198065

See also

Author

Mahmood Ul Hassan (scenic555@gmail.com); Frank Miller (frank.miller@liu.se)

Examples

# 2PL-models with discriminat and difficulty parameters
ip <- cbind(c(1.6, 1.6),c(-1, 1))

yyy <- optical(ip, oc="D", uncert=FALSE, ipop,
               imf=c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
               maxiter=rep(300, 6), eps=rep(0.002, 6),
               nnn=c(0, 50, 50, 200, 200, 200),
               nsp=c(0.001, 0.0001, 0.0001, 0.00001, 0.00001, 0.00001),
               sss=0.001, falpha=1.08, ig=3, ex=0)
#> ---> Outer iteration = 1 
#> ++++++++++++++++++
#> ---> Adapt grid; outer iteration = 2 
#> ++

yyy$h1
#>            [,1]      [,2] [,3] [,4]
#> bound  -0.73445 0.2313373    1    2
#> bound   0.00005 0.2686827    2    1
#> bound   0.73445 0.2686428    1    2
#> bound 100.00000 0.2313373    2    2

# 1PL-models with common discrimination parameter
ip <- cbind(c(1.6, NA), c(-1, 1))

yyy <- optical(ip, oc="D", uncert=FALSE, ipop,
               imf=c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
               maxiter=rep(300, 6), eps=rep(0.002, 6),
               nnn=c(0, 50, 50, 200, 200, 200),
               nsp=c(0.001, 0.0001, 0.0001, 0.00001, 0.00001, 0.00001),
               sss=0.001, falpha=1.08, ig=3, ex=0)
#> ---> Outer iteration = 1 
#> ++++++++++++++++++++
#> ---> Adapt grid; outer iteration = 2 
#> ++

yyy$h1
#>            [,1]      [,2] [,3] [,4]
#> bound  -0.44545 0.3279973    1    2
#> bound  -0.00005 0.1719828    2    1
#> bound   0.44585 0.1721671    1    2
#> bound 100.00000 0.3278528    2    2